move utils to utils
