+++ frontend/src/components/Chat/AddDm.vue
@@ -14,7 +14,7 @@
 </template>
 
 <script setup lang="ts">
-import { socket } from '../../socket';
+import { socket } from '../../utils/socket';
 import axiosInstance from '../../axiosConfig';
 import { ref } from 'vue';
 import Dialog from 'primevue/dialog';
