+++ frontend/src/components/Chat/AddMessage.vue
@@ -8,7 +8,7 @@
 </template>
 
 <script setup lang="ts">
-import { socket } from '../../socket';
+import { socket } from '../../utils/socket';
 import { ref } from 'vue';
 import { useToast } from 'primevue/usetoast';
