+++ frontend/src/components/Chat/ChannelDisplay.vue
@@ -13,7 +13,7 @@
 </template>
 
 <script setup lang="ts">
-import { socket } from '../../socket';
+import { socket } from '../../utils/socket';
 import axiosInstance from '../../axiosConfig';
 import { onBeforeMount, ref} from 'vue'
