+++ frontend/src/components/Friends/FriendsMenubar/FriendsAchievements.vue
@@ -17,7 +17,7 @@
 
 <script lang="ts">
 import { onBeforeMount, ref, watch } from 'vue';
-import axiosInstance from '../../../axiosConfig';
+import axiosInstance from '../../../utils/axiosConfig'
 
 export default {
   props: {
