+++ frontend/src/components/pong/Multiplayer.vue
@@ -21,11 +21,11 @@
 </template>
 
 <script lang="ts">
-	import { socket } from '../../socket'
+	import { socket } from '../../utils/socket'
 	import { defineComponent } from 'vue'
 	import {socket_match_id, p1_socket_id, p2_socket_id, username1, username2, match_id} from './MatchMaking.vue'
 	import { useRouter } from 'vue-router'
-	import axiosInstance from '../../axiosConfig'
+	import axiosInstance from '../../utils/axiosConfig'
 
 	import GameTools from './GameTools.vue'
 	import ResultScreen from './ResultScreen.vue'
